team members shown on team

+++ app/helpers/teams_helper.rb
@@ -11,11 +11,15 @@ module TeamsHelper
     end
   end
 
+  def leave_team_btn(tournament)
+    if tournament.organizer != current_user
+      link_to "Leave Team", leave_profile_team_path(current_user.team_for_tournament(@tournament)), data: { confirm: 'Are you sure?','sweet-alert-type': 'warning', text: 'You are about to leave this team & tournament', 'confirm-button-color': '#EE543A' }, method: :delete, class: "btn btn-danger btn-xs round"
+    end
+  end
+
   def leave_tournament_btn(tournament, team)
-		if user_signed_in?
-			if team.users.exists?(current_user) && team.captain == current_user
-				link_to "Leave Tournament", profile_tournament_team_path(tournament, team), method: :delete, data: { confirm: 'Are you sure?','sweet-alert-type': 'warning', text: 'You are about to remove your team from this tournament', 'confirm-button-color': '#EE543A' }, class: "btn btn-danger btn-xs round"
-			end
+		if team.users.exists?(current_user) && tournament.organizer != current_user && team.captain == current_user
+			link_to "Leave Tournament", profile_tournament_team_path(tournament, team), method: :delete, data: { confirm: 'Are you sure?','sweet-alert-type': 'warning', text: 'You are about to remove your team from this tournament', 'confirm-button-color': '#EE543A' }, class: "btn btn-danger btn-xs round"
 		end
 	end
